requirements.txt
openai
python-dotenv
html2text
markdown
unidiff
PyGithub
logging
pytest
pytest-mock
ossaudiodev
